Bug fixes and fix API calls to ensure compliance with Nominatim API rules.
Primary fix was to use node-fetch instead of axios and to set the user agent. Additionally, use the viewbox parameter on the search endpoint to have Nominatim prefer to return results in that area. It's not perfect and doing a manual calculation would be better, but it should do, one way or another.